Major refactoring. Pickling does not work correctly atm. (see sin.py)

+++ cinterpol/newton_interval.h
@@ -1,7 +1,6 @@
 #ifndef _NEWTON_INTERVAL_H_
 #define _NEWTON_INTERVAL_H_
 #include <math.h>    /* sqrt() round() */
-#include <stdlib.h>  /* size_t */
 #include <stdbool.h> /* bool type */
-size_t get_interval(const double arr[], const size_t N, const double t);
+int get_interval(const double arr[], const int N, const double t);
 #endif /* _NEWTON_INTERVAL_H_ */
